+++ engine/components/print_details.py
@@ -1,5 +1,5 @@
 """ Insert """
-from Logging.logging_config import logger
+from engine_logging.logging_config import logger
 
 
 def print_details(name: str, *args):
