+++ engine/computation/table_matrix/display_matrix.py
@@ -1,5 +1,5 @@
 """ Insert """
-from Logging.logging_config import logger
+from engine_logging.logging_config import logger
 
 
 def display_matrix(number_of_variables: int, returned_matrix: list, all_variables: list) -> None:
