+++ engine/computation/table_matrix/recursive_table.py
@@ -1,6 +1,6 @@
 """ Insert """
-from Components.constants import TRUE, FALSE
-from Logging.logging_config import logger
+from components.constants import TRUE, FALSE
+from engine_logging.logging_config import logger
 
 
 def recursive_table(n):
