+++ engine/main.py
@@ -1,8 +1,8 @@
 """
 This is the main entry for Engine. 
 """
-from Logging.logging_config import logger
-from Simulator.simulate_main import simulate_main
+from engine_logging.logging_config import logger
+from simulator.simulate_main import simulate_main
 
 
 logger.info("Starting Engine...")
