+++ engine/simulator/simulate_server.py
@@ -1,9 +1,9 @@
 """
 This can be used to simulate the calls that are made by the GRPC Server
 """
-from Parser.Checks import check_brackets, check_for_illegal_characters
-from Parser.Checks.check_brackets import UnequalBracketsExcept
-from Parser.Checks.check_for_illegal_characters import IllegalCharactersException
+from parser.checks import check_brackets, check_for_illegal_characters
+from parser.checks.check_brackets import UnequalBracketsExcept
+from parser.checks.check_for_illegal_characters import IllegalCharactersException
 # from input_statements import statement
